Move MeterView and HistoryView subviews
