+++ NammaMeter/HistoryView.swift
@@ -129,232 +129,6 @@ private struct HistoryContentView: View {
   }
 }
 
-private struct HistoryToolbarTitle: View {
-  var body: some View {
-    VStack(spacing: 2) {
-      Text("Trips")
-        .font(.nammaDisplay(18))
-      Text("ಪ್ರಯಾಣಗಳು")
-        .font(.nammaBody(12))
-    }
-  }
-}
-
-private struct HistorySearchBar: View {
-  @Binding var searchText: String
-  @FocusState.Binding var searchFocused: Bool
-
-  var body: some View {
-    HStack(spacing: 12) {
-      Image(systemName: "magnifyingglass")
-        .foregroundStyle(Color(uiColor: .secondaryLabel))
-      TextField("Search trips · ಹುಡುಕಿ", text: $searchText)
-        .font(.system(size: 16))
-        .foregroundStyle(Color(uiColor: .label))
-        .textInputAutocapitalization(.never)
-        .disableAutocorrection(true)
-        .focused($searchFocused)
-        .submitLabel(.search)
-      if !searchText.isEmpty {
-        Button {
-          searchText = ""
-          searchFocused = false
-        } label: {
-          Image(systemName: "xmark.circle.fill")
-            .foregroundStyle(Color(uiColor: .tertiaryLabel))
-        }
-        .buttonStyle(.plain)
-        .accessibilityLabel("Clear search")
-      } else if searchFocused {
-        Button {
-          searchFocused = false
-        } label: {
-          Image(systemName: "xmark.circle.fill")
-            .foregroundStyle(Color(uiColor: .tertiaryLabel))
-        }
-        .buttonStyle(.plain)
-        .accessibilityLabel("Cancel search")
-      }
-    }
-    .padding(.horizontal, 10)
-    .padding(.vertical, 8)
-    .background(Color(uiColor: .systemGray6))
-    .overlay(
-      RoundedRectangle(cornerRadius: 10, style: .continuous)
-        .stroke(Color(uiColor: .systemGray4), lineWidth: 0.5)
-    )
-    .clipShape(RoundedRectangle(cornerRadius: 10, style: .continuous))
-  }
-}
-
-private struct HistoryActionBar: View {
-  let isEditing: Bool
-  let showSelectAll: Bool
-  let showDelete: Bool
-  let isAllSelected: Bool
-  let onToggleSelectAll: () -> Void
-  let onDeleteSelected: () -> Void
-  let onToggleEdit: () -> Void
-
-  var body: some View {
-    HStack(spacing: 12) {
-      if showSelectAll {
-        Button {
-          onToggleSelectAll()
-        } label: {
-          mangoPillLabel(isAllSelected ? "Deselect All" : "Select All")
-        }
-        .accessibilityLabel(isAllSelected ? "Deselect All" : "Select All")
-      }
-      Spacer()
-      if showDelete {
-        Button(role: .destructive) {
-          onDeleteSelected()
-        } label: {
-          Label("Delete", systemImage: "trash")
-        }
-        .tint(.red)
-        .buttonStyle(.bordered)
-        .controlSize(.small)
-      }
-      Button {
-        onToggleEdit()
-      } label: {
-        mangoPillLabel(isEditing ? "Done" : "Edit")
-      }
-      .accessibilityLabel(isEditing ? "Done" : "Edit")
-    }
-  }
-
-  private func mangoPillLabel(_ title: String) -> some View {
-    Text(title)
-      .font(.nammaDisplay(13))
-      .foregroundStyle(Theme.ink)
-      .padding(.horizontal, 14)
-      .padding(.vertical, 6)
-      .background(Theme.mango.opacity(0.6))
-      .clipShape(Capsule())
-  }
-}
-
-private struct HistoryListContent: View {
-  let trips: [Trip]
-  let filteredTrips: [Trip]
-  let isEditing: Bool
-  let onDeleteFiltered: (IndexSet) -> Void
-
-  @ViewBuilder
-  var body: some View {
-    if trips.isEmpty {
-      emptyState
-    } else if filteredTrips.isEmpty {
-      noResultsState
-    } else {
-      tripRows
-    }
-  }
-
-  private var emptyState: some View {
-    VStack(alignment: .center, spacing: 12) {
-      Text("No trips yet")
-        .font(.nammaDisplay(16))
-        .foregroundStyle(Theme.ink)
-      Text("ಯಾವುದೇ ಪ್ರಯಾಣಗಳಿಲ್ಲ")
-        .font(.nammaBody(12))
-        .foregroundStyle(Theme.ink.opacity(0.7))
-    }
-    .frame(maxWidth: .infinity, minHeight: 120)
-    .listRowBackground(Color.clear)
-  }
-
-  private var noResultsState: some View {
-    VStack(alignment: .center, spacing: 12) {
-      Text("No matching trips")
-        .font(.nammaDisplay(16))
-        .foregroundStyle(Theme.ink)
-      Text("ಹೊಂದುವ ಪ್ರಯಾಣಗಳಿಲ್ಲ")
-        .font(.nammaBody(12))
-        .foregroundStyle(Theme.ink.opacity(0.7))
-    }
-    .frame(maxWidth: .infinity, minHeight: 120)
-    .listRowBackground(Color.clear)
-  }
-
-  private var tripRows: some View {
-    ForEach(filteredTrips) { trip in
-      tripRow(for: trip)
-    }
-    .onDelete(perform: onDeleteFiltered)
-  }
-
-  @ViewBuilder
-  private func tripRow(for trip: Trip) -> some View {
-    if isEditing {
-      TripRow(trip: trip)
-        .tag(trip.id)
-        .listRowBackground(Theme.card)
-    } else {
-      NavigationLink {
-        TripDetailView(tripId: trip.id)
-      } label: {
-        TripRow(trip: trip)
-      }
-      .tag(trip.id)
-      .listRowBackground(Theme.card)
-    }
-  }
-}
-
-struct TripRow: View {
-  let trip: Trip
-
-  var body: some View {
-    VStack(alignment: .leading, spacing: 8) {
-      HStack {
-        if let name = trip.name, !name.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty {
-          Text(name)
-            .font(.nammaDisplay(14))
-        } else {
-          Text(trip.startDate, format: .dateTime.day().month().hour().minute())
-            .font(.nammaDisplay(14))
-        }
-        Spacer()
-        Text(trip.fare, format: .currency(code: "INR"))
-          .font(.nammaDisplay(14))
-      }
-
-      if let name = trip.name, !name.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty {
-        Text(trip.startDate, format: .dateTime.day().month().year().hour().minute())
-          .font(.nammaBody(11))
-          .foregroundStyle(Theme.ink.opacity(0.7))
-      }
-
-      HStack(spacing: 12) {
-        Label {
-          Text(trip.startLocationName ?? "Locating...")
-        } icon: {
-          Image(systemName: "mappin.and.ellipse")
-        }
-
-        Label {
-          Text(formattedElapsed(trip.duration))
-        } icon: {
-          Image(systemName: "clock")
-        }
-
-        Label {
-          Text("\((trip.distanceMeters / 1000).formatted(.number.precision(.fractionLength(2)))) km")
-        } icon: {
-          Image(systemName: "map")
-        }
-      }
-      .font(.nammaBody(11))
-      .foregroundStyle(Theme.ink.opacity(0.7))
-    }
-    .padding(.vertical, 6)
-  }
-}
-
 #Preview {
   HistoryView()
     .environment(TripStore())
