+++ NammaMeter/Views/History/HistoryToolbarTitle.swift
@@ -1,0 +1,12 @@
+import SwiftUI
+
+struct HistoryToolbarTitle: View {
+  var body: some View {
+    VStack(spacing: 2) {
+      Text("Trips")
+        .font(.nammaDisplay(18))
+      Text("ಪ್ರಯಾಣಗಳು")
+        .font(.nammaBody(12))
+    }
+  }
+}
